error debug info verified

+++ hw2/decafch.py
@@ -6,6 +6,7 @@ from decafparser import decaf_parser
 arg_parser = argparse.ArgumentParser(description='Decaf language checker')
 arg_parser.add_argument('filename', help='file to compile')
 arg_parser.add_argument('-d', action='store_true', help='enable debug mode')
+arg_parser.add_argument('-c', action='store_true', help='clean up intermediate files')
 args = arg_parser.parse_args()
 
 try:
@@ -16,7 +17,8 @@ try:
         print "empty file"
     else:
         decaf_parser.parse(inbuf, debug=args.d)
-        os.system("rm *.pyc");
+        if args.c:
+            os.system("rm *.pyc");
 
 except IOError:
     print args.filename+ ' not exist'
